+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Changed typography sizes and added breakpoint styling for headings/small.
 - Changed topbar logo to be an image wrapped by an anchor to be more accessible.
+- Renamed .expandable-content to .expandable-body.
 
 ### Fixed